Vorbereitung für andere Karten
