+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Die Karte wurde etwas generalisiert und ermöglicht jetzt leichter, auch andere Daten zu zeigen.
+
+### Fixed
+- Das Label für Hochschulstandorte zeigt jetzt an, dass es sich um ORCA-Mitglieder handelt, die nur keine Open Content Policy haben.
+- Der Leaflet-Container füllt nun die komplette Größe des Containers.
+
 ## [0.4.0] - 2022-09-01
 ### Changed
 - Das initiale Zoomlevel hängt nun von der Größe des Containers ab. Ist der Container größer als 524 px in mindestens einer Dimension, ist das Zoomlevel 1, sonst 0.
