+++ beratungsstellen.html
@@ -16,7 +16,9 @@
 
 <body>
   <div class="leaflet-container" id="map"></div>
-  <script type="module" src="scripts/main.js"></script>
+  <script type="module" src="scripts/define-markers.js"></script>
+  <script type="module" src="scripts/create-map.js"></script>
+  <script type="module" src="scripts/beratungsstellen.js"></script>
 </body>
 
 </html>
