+++ css/style.css
@@ -36,10 +36,8 @@ code {
 }
 
 .leaflet-container {
-  height: 1024px;
-  width: 1024px;
-  max-width: 100%;
-  max-height: 100%;
+  height: 100%;
+  width: 100%;
 }
 
 .leaflet-popup-content {
