--- db/liste.md
@@ -1,57 +1,0 @@
-Bergische Universität Wuppertal x
-Fachhochschule Bielefeld x 400x385
-Fachhochschule Dortmund x
-Fachhochschule Münster x
-Heinrich Heine Universität Düsseldorf x
-Hochschule Bonn-Rhein-Sieg, Campus Rheinbach x
-Hochschule Bonn-Rhein-Sieg, Campus Sankt Augustin x
-Hochschule Düsseldorf x
-Hochschule Hamm-Lippstadt, Campus Hamm x
-Hochschule Hamm-Lippstadt - Campus Lippstadt x
-Ruhr-Universität Bochum x
-RWTH Aachen University x
-Technische Hochschule Köln, Campus Südstadt x
-Technische Hochschule Ostwestfalen-Lippe, Campus Detmold x
-Technische Hochschule Ostwestfalen-Lippe, Campus Höxter x
-Technische Hochschule Ostwestfalen-Lippe, Campus Lemgo x
-Universität Bielefeld x
-Universität zu Köln x
-Universität Paderborn x
-Universität Siegen x
-Westfälische Hochschule, Campus Recklinghausen x
-Westfälische Wilhelms-Universität Münster x
-Westfälische Hochschule, Campus Gelsenkirchen x
-Hochschule Bochum, Campus Bochum x
-Fachhochschule Südwestfalen, Campus Hagen x
-Fachhochschule Südwestfalen, Campus Iserlohn x
-Fachhochschule Südwestfalen, Campus Meschede x
-Fachhochschule Südwestfalen, Campus Lüdenscheid x
-Fachhochschule Südwestfalen, Campus Soest x
-Hochschule Niederrhein, Campus Mönchengladbach x
-Hochschule Niederrhein, Campus Krefeld x
-Rheinische Friedrich-Wilhelms-Universität Bonn x
-Universität Duisburg-Essen, Campus Essen x
-Universität Duisburg-Essen, Campus Duisburg x
-Fachhochschule Aachen x
-Hochschule Rhein-Waal, Campus Kleve x
-Hochschule Rhein-Waal, Campus Kamp-Lintfort x
-Hochschule Ruhr West, Campus Mülheim x
-Hochschule Ruhr West, Campus Bottrop x
-Hochschule für Gesundheit Bochum x
-Hochschule für Musik und Tanz Köln x
-Hochschule für Musik Detmold x
-Folkwang Universität der Künste, Campus Essen x
-Deutsche Sporthochschule Köln x
-Technische Universität Dortmund x
-FernUniversität in Hagen x
-Robert Schumann Hochschule Düsseldorf x
-Kunsthochschule für Medien Köln x
-Kunstakademie Münster x
-Kunstakademie Düsseldorf x
-Westfälische Hochschule Campus Bocholt x
-
-52,53075 N, 9,46167 E (rechtsoben)
-
-50,32381 N, 5,86776 E (linksunten)
-
-2,20694 x 3,59391
