+++ index.html
@@ -13,11 +13,11 @@
 <body>
   <main>
     <h1>OER Policies an NRW Hochschulen</h1>
-    <center><iframe src="karte.html" height="524" width="524"></iframe></center>
+    <center><iframe src="netzwerk.html" height="524" width="524"></iframe></center>
     <p></p>
     <details>
       <summary>Einbettungscode</summary>
-      <p><code>&lt;iframe src="https://mrkwnzl.github.io/orca-nrw-karte/karte.html" height="524" width="524"&gt;&lt;/iframe&gt;</code></p>
+      <p><code>&lt;iframe src="https://mrkwnzl.github.io/orca-nrw-karte/netzwerk.html" height="524" width="524"&gt;&lt;/iframe&gt;</code></p>
     </details>
   </main>
   <footer>
