+++ scripts/define-markers.js
@@ -1,0 +1,28 @@
+// Marker definieren
+export const blauerPunkt = {
+  color: 'rgb(42,168,226)',
+  fillColor: 'rgb(42,168,226)',
+  fillOpacity: 1,
+  radius: 2
+};
+
+export const orangerPunkt = {
+  color: 'rgb(249,147,28)',
+  fillColor: 'rgb(249,147,28)',
+  fillOpacity: 1,
+  radius: 2
+};
+
+export const gruenerPunkt = {
+  color: 'rgb(0,104,55)',
+  fillColor: 'rgb(0,104,55)',
+  fillOpacity: 1,
+  radius: 2
+}
+
+export const grauerPunkt = {
+  color: 'rgb(128,128,128)',
+  fillColor: 'rgb(128,128,128)',
+  fillOpacity: 1,
+  radius: 1.5
+}
